+++ app/src/main/java/com/example/pawdopt/data/model/User.kt
@@ -1,7 +1,7 @@
 package com.example.pawdopt.data.model
 
-data class User (
-    val id: Int = 0,
+data class User(
+    val id: Long = 0,
     val nombre: String,
     val email: String,
     val password: String,
